Add asset caching
In order to overcome the peformance issues that NFS is generating
asset caching will behave similarily to how needle caching is done
this will also pave the way to implement proper caching later on.

Asset caching is enabled when CACHEASSET is set on workers.ini and
CACHEDIRECTORY is set (Therefore bound to needle caching for now)

Concept is simple: If caching is enabled, and asset caching is enabled
too, then lock the file, rsync it, and replace variables as needed for the
backend to continue it's work.